tests/test_basic.py
"""
Basic tests for RAG implementation.

Run with: pytest tests/test_basic.py -v
"""

import pytest
import torch
import sys
import os

# Add parent directory to path
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from rag import RAGSequenceForGeneration, RAGTokenForGeneration, RAGConfig
from rag.retrieval import MockRetriever, DPRRetriever, BM25Retriever
from rag.generation import BARTGenerator
from rag.config import get_default_config


class TestConfig:
    """Test configuration classes."""

    def test_rag_config_creation(self):
        """Test RAG config creation."""
        config = RAGConfig()
        assert config.model_type in ["rag_sequence", "rag_token"]
        assert config.num_retrieved_docs > 0
        assert config.generator_max_length > 0

    def test_rag_config_save_load(self, tmp_path):
        """Test config save and load."""
        config = RAGConfig(num_retrieved_docs=10)

        # Save
        config_path = tmp_path / "config.json"
        config.save(str(config_path))

        # Load
        loaded_config = RAGConfig.load(str(config_path))
        assert loaded_config.num_retrieved_docs == 10

    def test_default_configs(self):
        """Test default configurations for different tasks."""
        tasks = ["open_qa", "abstractive_qa", "question_generation", "fact_verification"]

        for task in tasks:
            config_dict = get_default_config(task)
            assert "rag" in config_dict
            assert "training" in config_dict
            assert "task" in config_dict


class TestRetrievers:
    """Test retriever components."""

    def test_mock_retriever(self):
        """Test mock retriever."""
        retriever = MockRetriever(num_docs=100, embed_dim=768)

        # Single query
        results = retriever.retrieve(["test query"], k=5)
        assert len(results) == 1
        assert len(results[0]) == 5
        assert all(hasattr(doc, "text") for doc in results[0])
        assert all(hasattr(doc, "score") for doc in results[0])

    def test_mock_retriever_batch(self):
        """Test mock retriever with batch."""
        retriever = MockRetriever(num_docs=100)

        queries = ["query 1", "query 2", "query 3"]
        results = retriever.retrieve(queries, k=5)

        assert len(results) == 3
        assert all(len(docs) == 5 for docs in results)

    def test_bm25_retriever(self):
        """Test BM25 retriever."""
        passages = [
            {"id": "1", "text": "Paris is the capital of France", "title": "Paris"},
            {"id": "2", "text": "London is the capital of UK", "title": "London"},
            {"id": "3", "text": "Berlin is the capital of Germany", "title": "Berlin"},
        ]

        retriever = BM25Retriever(passages=passages)

        # Retrieve
        results = retriever.retrieve(["capital of France"], k=2)

        assert len(results) == 1
        assert len(results[0]) == 2
        # First result should be about Paris
        assert "Paris" in results[0][0].text or "France" in results[0][0].text


class TestGenerator:
    """Test generator component."""

    @pytest.mark.slow
    def test_bart_generator_loading(self):
        """Test BART generator loading."""
        # Use small model for testing
        generator = BARTGenerator(model_name="facebook/bart-base")

        assert generator.model is not None
        assert generator.tokenizer is not None

    def test_bart_generator_prepare_inputs(self):
        """Test input preparation."""
        generator = BARTGenerator(model_name="facebook/bart-base")

        inputs = generator.prepare_inputs(
            query="What is the capital?",
            context="Paris is the capital of France.",
            max_length=128,
        )

        assert "input_ids" in inputs
        assert "attention_mask" in inputs
        assert inputs["input_ids"].shape[1] <= 128


class TestRAGModels:
    """Test RAG models."""

    @pytest.fixture
    def mock_rag_sequence(self):
        """Create RAG-Sequence model with mock retriever."""
        config = RAGConfig(
            model_type="rag_sequence",
            num_retrieved_docs=3,
            generator_name_or_path="facebook/bart-base",
        )
        retriever = MockRetriever(num_docs=50)
        model = RAGSequenceForGeneration(config=config, retriever=retriever)
        return model

    @pytest.fixture
    def mock_rag_token(self):
        """Create RAG-Token model with mock retriever."""
        config = RAGConfig(
            model_type="rag_token",
            num_retrieved_docs=3,
            generator_name_or_path="facebook/bart-base",
        )
        retriever = MockRetriever(num_docs=50)
        model = RAGTokenForGeneration(config=config, retriever=retriever)
        return model

    @pytest.mark.slow
    def test_rag_sequence_initialization(self, mock_rag_sequence):
        """Test RAG-Sequence initialization."""
        assert mock_rag_sequence.retriever is not None
        assert mock_rag_sequence.generator is not None
        assert mock_rag_sequence.num_docs == 3

    @pytest.mark.slow
    def test_rag_token_initialization(self, mock_rag_token):
        """Test RAG-Token initialization."""
        assert mock_rag_token.retriever is not None
        assert mock_rag_token.generator is not None
        assert mock_rag_token.num_docs == 3

    @pytest.mark.slow
    def test_rag_sequence_forward(self, mock_rag_sequence):
        """Test RAG-Sequence forward pass."""
        # Create dummy inputs
        input_ids = torch.randint(0, 1000, (2, 10))  # batch=2, seq_len=10
        labels = torch.randint(0, 1000, (2, 8))  # target_len=8

        # Forward pass
        outputs = mock_rag_sequence(
            input_ids=input_ids,
            labels=labels,
        )

        assert "loss" in outputs
        assert outputs["loss"].item() >= 0  # Loss should be non-negative

    @pytest.mark.slow
    def test_rag_token_forward(self, mock_rag_token):
        """Test RAG-Token forward pass."""
        # Create dummy inputs
        input_ids = torch.randint(0, 1000, (2, 10))
        labels = torch.randint(0, 1000, (2, 8))

        # Forward pass
        outputs = mock_rag_token(
            input_ids=input_ids,
            labels=labels,
        )

        assert "loss" in outputs
        assert outputs["loss"].item() >= 0

    @pytest.mark.slow
    @pytest.mark.integration
    def test_rag_sequence_generate(self, mock_rag_sequence):
        """Test RAG-Sequence generation."""
        # Tokenize a query
        query = "What is the capital of France?"
        inputs = mock_rag_sequence.generator.tokenizer(
            query,
            return_tensors="pt",
            max_length=64,
            truncation=True,
        )

        # Generate
        output_ids = mock_rag_sequence.generate(
            input_ids=inputs["input_ids"],
            attention_mask=inputs["attention_mask"],
            max_length=20,
            num_beams=2,
        )

        assert output_ids is not None
        assert output_ids.shape[0] == 1  # batch size

    @pytest.mark.slow
    @pytest.mark.integration
    def test_rag_generate_from_query(self, mock_rag_sequence):
        """Test generate_from_query convenience method."""
        # This might fail with mock retriever, but should not crash
        try:
            answer = mock_rag_sequence.generate_from_query(
                query="What is RAG?",
                max_length=20,
                num_beams=1,
            )
            assert isinstance(answer, str)
        except Exception as e:
            # Expected with mock retriever
            print(f"Note: Generation with mock retriever raised: {e}")


class TestEndToEnd:
    """End-to-end integration tests."""

    @pytest.mark.slow
    @pytest.mark.integration
    def test_simple_qa_pipeline(self):
        """Test simple QA pipeline."""
        # Create sample documents
        passages = [
            {"id": "1", "text": "Paris is the capital of France.", "title": "Paris"},
            {"id": "2", "text": "London is the capital of UK.", "title": "London"},
        ]

        # Create BM25 retriever
        retriever = BM25Retriever(passages=passages)

        # Create RAG model
        config = RAGConfig(
            num_retrieved_docs=2,
            generator_name_or_path="facebook/bart-base",
        )
        model = RAGSequenceForGeneration(config=config, retriever=retriever)

        # Test retrieval
        results = retriever.retrieve(["capital of France"], k=2)
        assert len(results) == 1
        assert len(results[0]) == 2


# Pytest configuration
def pytest_configure(config):
    """Configure pytest."""
    config.addinivalue_line(
        "markers", "slow: marks tests as slow (deselect with '-m \"not slow\"')"
    )
    config.addinivalue_line(
        "markers", "integration: marks tests as integration tests"
    )


if __name__ == "__main__":
    # Run basic tests
    pytest.main([__file__, "-v", "-m", "not slow"])
