requirements.txt
# Core Dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
accelerate>=0.20.0

# Retrieval & Indexing
faiss-gpu>=1.7.4  # Use faiss-cpu if no GPU available
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
nltk>=3.8.1

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
wandb>=0.15.0  # For experiment tracking
tensorboard>=2.13.0

# Evaluation
rouge-score>=0.1.2
sacrebleu>=2.3.0
scikit-learn>=1.3.0

# Optional: For Wikipedia processing
mwparserfromhell>=0.6.4
beautifulsoup4>=4.12.0
requests>=2.31.0
